requirements.txt
pandas
requests
python-dotenv
db-sqlite3